requirements.txt
flask
gunicorn
psycopg2-binary
python-dotenv
httpx
beautifulsoup4
sentence-transformers